deploy/pyproject.toml
[tool.poetry]
name = "deploy"
version = "0.1.0"
description = ""
authors = ["Kshitiz <dzgrotechnologies@gmail.com>"]
readme = "README.md"
packages = [{ include = "deploy", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
inquirer = "*"
toml = "*"
pyyaml = "6.0.2"
psutil = "7.0.0"
python-dotenv = "1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
 
